+++ ConCurrency/ExchangeCurrency.hpp
@@ -13,8 +13,8 @@ public:
 	ExchangeCurrency(std::shared_ptr<CashRegister> cashRegister);
 
 	virtual ~ExchangeCurrency() = default;
-	virtual std::pair<int, double> currencyExchange(int amount, std::string currencyCode) = 0;
+	virtual std::pair<double, double> currencyExchange(double amount, std::string currencyCode) = 0;
 
-	bool checkCurrencyAvailability(std::pair<int, double> exchangedMoney, std::string currencyCode);
-	void updateCurrenciesAmounts(int amount, std::pair<int, double> exchangedMoney, std::string currencyCode);
+	bool checkCurrencyAvailability(std::pair<double, double> exchangedMoney, std::string currencyCode);
+	void updateCurrenciesAmounts(double amount, std::pair<double, double> exchangedMoney, std::string currencyCode);
 };
